refactor(link): split the HarmonyOS native-object scan out of build_and_run.rs
build_and_run.rs was at 1994 lines; #8930's archive-group call takes it over
the 2000-line gate. The HarmonyOS block needs only cmd/target/format, so it
moves to a sibling module unchanged.
